requirements.txt
torch==1.7.0 
torchvision==0.8.1
mmcv-full==1.3.3
opencv-python>=4.4.0
texttable==1.6.4
terminaltables==3.1.0
pycocotools==2.0.2
visdom
Pillow
tqdm